23 Beaconside
"Thank you. That is quite settled," said he, rising and putting his lens in his pocket. "Hullo! Here
is something interesting!"

The object which had caught his eye was a small dog lash hung on one corner of the bed. The lash,
however, was curled upon itself and tied so as to make a loop of whipcord.

"What do you make of that, Watson?"

"It's a common enough lash. But I don't know why it should be tied."

"That is not quite so common, is it? Ah, me! it's a wicked world, and when a clever man turns his
brains to crime it is the worst of all. I think that I have seen enough now, Miss Stoner, and with
your permission we shall walk out upon the lawn."