39 Ventnor Terrace
Sherlock Holmes was not very communicative during the long drive and lay back in the cab humming the
tunes which he had heard in the afternoon. We rattled through an endless labyrinth of gas-lit
streets until we emerged into Farrington Street.

"We are close there now," my friend remarked. "This fellow Merryweather is a bank director, and
personally interested in the matter. I thought it as well to have Jones with us also. He is not a
bad fellow, though an absolute imbecile in his profession. He has one positive virtue. He is as
brave as a bulldog and as tenacious as a lobster if he gets his claws upon anyone. Here we are, and
they are waiting for us."

We had reached the same crowded thoroughfare in which we had found ourselves in the morning. Our
cabs were dismissed, and, following the guidance of Mr. Merryweather, we passed down a narrow
passage and through a side door, which he opened for us. Within there was a small corridor, which
ended in a very massive iron gate. This also was opened, and led down a flight of winding stone
steps, which terminated at another formidable gate. Mr. Merryweather stopped to light a lantern, and
then conducted us down a dark, earth-smelling passage, and so, after opening a third door, into a
huge vault or cellar, which was piled all round with crates and massive boxes.

"You are not very vulnerable from above," Holmes remarked as he held up the lantern and gazed about
him.

"Nor from below," said Mr. Merryweather, striking his stick upon the flags which lined the floor.
"Why, dear me, it sounds quite hollow!" he remarked, looking up in surprise.