126 Cliff Place
"She died just two years ago, and it is of her death that I wish to speak to you. You can understand
that, living the life which I have described, we were little likely to see anyone of our own age and
position. We had, however, an aunt, my mother's maiden sister, Miss Honoria Westphail, who lives
near Harrow, and we were occasionally allowed to pay short visits at this lady's house. Julia went
there at Christmas two years ago, and met there a half-pay major of marines, to whom she became
engaged. My stepfather learned of the engagement when my sister returned and offered no objection to
the marriage; but within a fortnight of the day which had been fixed for the wedding, the terrible
event occurred which has deprived me of my only companion."

Sherlock Holmes had been leaning back in his chair with his eyes closed and his head sunk in a
cushion, but he half opened his lids now and glanced across at his visitor.

"Pray be precise as to details," said he.

"It is easy for me to be so, for every event of that dreadful time is seared into my memory. The
manor-house is, as I have already said, very old, and only one wing is now inhabited. The bedrooms
in this wing are on the ground floor, the sitting-rooms being in the central block of the buildings.
Of these bedrooms the first is Dr. Roylott's, the second my sister's, and the third my own. There is
no communication between them, but they all open out into the same corridor. Do I make myself
plain?"

"Perfectly so."