110 Larch Walk
"Do not go asleep; your very life may depend upon it. Have your pistol ready in case we should need
it. I will sit on the side of the bed, and you in that chair."

I took out my revolver and laid it on the corner of the table.

Holmes had brought up a long thin cane, and this he placed upon the bed beside him. By it he laid
the box of matches and the stump of a candle. Then he turned down the lamp, and we were left in
darkness.

How shall I ever forget that dreadful vigil? I could not hear a sound, not even the drawing of a
breath, and yet I knew that my companion sat open-eyed, within a few feet of me, in the same state
of nervous tension in which I was myself. The shutters cut off the least ray of light, and we waited
in absolute darkness.

From outside came the occasional cry of a night-bird, and once at our very window a long drawn
catlike whine, which told us that the cheetah was indeed at liberty. Far away we could hear the deep
tones of the parish clock, which boomed out every quarter of an hour. How long they seemed, those
quarters! Twelve struck, and one and two and three, and still we sat waiting silently for whatever
might befall.