19 Harvey Street
Sherlock Holmes was transformed when he was hot upon such a scent as this. Men who had only known
the quiet thinker and logician of Baker Street would have failed to recognise him. His face flushed
and darkened. His brows were drawn into two hard black lines, while his eyes shone out from beneath
them with a steely glitter. His face was bent downward, his shoulders bowed, his lips compressed,
and the veins stood out like whipcord in his long, sinewy neck. His nostrils seemed to dilate with a
purely animal lust for the chase, and his mind was so absolutely concentrated upon the matter before
him that a question or remark fell unheeded upon his ears, or, at the most, only provoked a quick,
impatient snarl in reply. Swiftly and silently he made his way along the track which ran through the
meadows, and so by way of the woods to the Boscombe Pool. It was damp, marshy ground, as is all that
district, and there were marks of many feet, both upon the path and amid the short grass which
bounded it on either side. Sometimes Holmes would hurry on, sometimes stop dead, and once he made
quite a little detour into the meadow. Lestrade and I walked behind him, the detective indifferent
and contemptuous, while I watched my friend with the interest which sprang from the conviction that
every one of his actions was directed towards a definite end.

The Boscombe Pool, which is a little reed-girt sheet of water some fifty yards across, is situated
at the boundary between the Hatherley Farm and the private park of the wealthy Mr. Turner. Above the
woods which lined it upon the farther side we could see the red, jutting pinnacles which marked the
site of the rich landowner's dwelling. On the Hatherley side of the pool the woods grew very thick,
and there was a narrow belt of sodden grass twenty paces across between the edge of the trees and
the reeds which lined the lake. Lestrade showed us the exact spot at which the body had been found,
and, indeed, so moist was the ground, that I could plainly see the traces which had been left by the
fall of the stricken man. To Holmes, as I could see by his eager face and peering eyes, very many
other things were to be read upon the trampled grass. He ran round, like a dog who is picking up a
scent, and then turned upon my companion.

"What did you go into the pool for?" he asked.

"I fished about with a rake. I thought there might be some weapon or other trace. But how on
earth--"