45 Down Close
"When I went down there I found him talking with his son, so I smoked a cigar and waited behind a
tree until he should be alone. But as I listened to his talk all that was black and bitter in me
seemed to come uppermost. He was urging his son to marry my daughter with as little regard for what
she might think as if she were a slut from off the streets. It drove me mad to think that I and all
that I held most dear should be in the power of such a man as this. Could I not snap the bond? I was
already a dying and a desperate man. Though clear of mind and fairly strong of limb, I knew that my
own fate was sealed. But my memory and my girl! Both could be saved if I could but silence that foul
tongue. I did it, Mr. Holmes. I would do it again. Deeply as I have sinned, I have led a life of
martyrdom to atone for it. But that my girl should be entangled in the same meshes which held me was
more than I could suffer. I struck him down with no more compunction than if he had been some foul
and venomous beast. His cry brought back his son; but I had gained the cover of the wood, though I
was forced to go back to fetch the cloak which I had dropped in my flight. That is the true story,
gentlemen, of all that occurred."

"Well, it is not for me to judge you," said Holmes as the old man signed the statement which had
been drawn out. "I pray that we may never be exposed to such a temptation."

"I pray not, sir. And what do you intend to do?"