23 Ventnor Terrace
"My God!" I whispered; "did you see it?"

Holmes was for the moment as startled as I. His hand closed like a vice upon my wrist in his
agitation. Then he broke into a low laugh and put his lips to my ear.

"It is a nice household," he murmured. "That is the baboon."

I had forgotten the strange pets which the doctor affected. There was a cheetah, too; perhaps we
might find it upon our shoulders at any moment. I confess that I felt easier in my mind when, after
following Holmes' example and slipping off my shoes, I found myself inside the bedroom. My companion
noiselessly closed the shutters, moved the lamp onto the table, and cast his eyes round the room.
All was as we had seen it in the daytime. Then creeping up to me and making a trumpet of his hand,
he whispered into my ear again so gently that it was all that I could do to distinguish the words:

"The least sound would be fatal to our plans."

I nodded to show that I had heard.