8 Wantage Close
Holmes chuckled and wriggled in his chair, as was his habit when in high spirits. "It is a little
off the beaten track, isn't it?" said he. "And now, Mr. Wilson, off you go at scratch and tell us
all about yourself, your household, and the effect which this advertisement had upon your fortunes.
You will first make a note, Doctor, of the paper and the date."

"It is The Morning Chronicle of April 27, 1890. Just two months ago."

"Very good. Now, Mr. Wilson?"