33 Beaconside
"Grave enough!" said Mr. Jabez Wilson. "Why, I have lost four pound a week."

"As far as you are personally concerned," remarked Holmes, "I do not see that you have any grievance
against this extraordinary league. On the contrary, you are, as I understand, richer by some £30, to
say nothing of the minute knowledge which you have gained on every subject which comes under the
letter A. You have lost nothing by them."

"No, sir. But I want to find out about them, and who they are, and what their object was in playing
this prank--if it was a prank--upon me. It was a pretty expensive joke for them, for it cost them
two and thirty pounds."

"We shall endeavour to clear up these points for you. And, first, one or two questions, Mr. Wilson.
This assistant of yours who first called your attention to the advertisement--how long had he been
with you?"