5 Tamworth Drive
"Now, then, mister," said he, with his head cocked and his arms akimbo, "what are you driving at?
Let's have it straight, now."

"It is straight enough. I should like to know who sold you the geese which you supplied to the
Alpha."

"Well then, I shan't tell you. So now!"

"Oh, it is a matter of no importance; but I don't know why you should be so warm over such a
trifle."

"Warm! You'd be as warm, maybe, if you were as pestered as I am. When I pay good money for a good
article there should be an end of the business; but it's 'Where are the geese?' and 'Who did you
sell the geese to?' and 'What will you take for the geese?' One would think they were the only geese
in the world, to hear the fuss that is made over them."