105 Larch Walk
"Are they not fresh and beautiful?" I cried with all the enthusiasm of a man fresh from the fogs of
Baker Street.

But Holmes shook his head gravely.

"Do you know, Watson," said he, "that it is one of the curses of a mind with a turn like mine that I
must look at everything with reference to my own special subject. You look at these scattered
houses, and you are impressed by their beauty. I look at them, and the only thought which comes to
me is a feeling of their isolation and of the impunity with which crime may be committed there."