22 Balcombe Close
"The band! the speckled band!" whispered Holmes.

I took a step forward. In an instant his strange headgear began to move, and there reared itself
from among his hair the squat diamond-shaped head and puffed neck of a loathsome serpent.

"It is a swamp adder!" cried Holmes; "the deadliest snake in India. He has died within ten seconds
of being bitten. Violence does, in truth, recoil upon the violent, and the schemer falls into the
pit which he digs for another. Let us thrust this creature back into its den, and we can then remove
Miss Stoner to some place of shelter and let the county police know what has happened."

As he spoke he drew the dog-whip swiftly from the dead man's lap, and throwing the noose round the
reptile's neck he drew it from its horrid perch and, carrying it at arm's length, threw it into the
iron safe, which he closed upon it.

Such are the true facts of the death of Dr. Grimesby Roylott, of Stoke Moran. It is not necessary
that I should prolong a narrative which has already run to too great a length by telling how we
broke the sad news to the terrified girl, how we conveyed her by the morning train to the care of
her good aunt at Harrow, of how the slow process of official inquiry came to the conclusion that the
doctor met his fate while indiscreetly playing with a dangerous pet. The little which I had yet to
learn of the case was told me by Sherlock Holmes as we travelled back next day.