10 Tamworth Drive
"My name," said he, "is John Openshaw, but my own affairs have, as far as I can understand, little
to do with this awful business. It is a hereditary matter; so in order to give you an idea of the
facts, I must go back to the commencement of the affair.

"You must know that my grandfather had two sons--my uncle Elias and my father Joseph. My father had
a small factory at Coventry, which he enlarged at the time of the invention of bicycling. He was a
patentee of the Openshaw unbreakable tire, and his business met with such success that he was able
to sell it and to retire upon a handsome competence.

"My uncle Elias emigrated to America when he was a young man and became a planter in Florida, where
he was reported to have done very well. At the time of the war he fought in Jackson's army, and
afterwards under Hood, where he rose to be a colonel. When Lee laid down his arms my uncle returned
to his plantation, where he remained for three or four years. About 1869 or 1870 he came back to
Europe and took a small estate in Sussex, near Horsham. He had made a very considerable fortune in
the States, and his reason for leaving them was his aversion to the negroes, and his dislike of the
Republican policy in extending the franchise to them. He was a singular man, fierce and quick-
tempered, very foul-mouthed when he was angry, and of a most retiring disposition. During all the
years that he lived at Horsham, I doubt if ever he set foot in the town. He had a garden and two or
three fields round his house, and there he would take his exercise, though very often for weeks on
end he would never leave his room. He drank a great deal of brandy and smoked very heavily, but he
would see no society and did not want any friends, not even his own brother.