77 Sunningdale Drive
"But a terrible change came over our stepfather about this time. Instead of making friends and
exchanging visits with our neighbours, who had at first been overjoyed to see a Roylott of Stoke
Moran back in the old family seat, he shut himself up in his house and seldom came out save to
indulge in ferocious quarrels with whoever might cross his path. Violence of temper approaching to
mania has been hereditary in the men of the family, and in my stepfather's case it had, I believe,
been intensified by his long residence in the tropics. A series of disgraceful brawls took place,
two of which ended in the police-court, until at last he became the terror of the village, and the
folks would fly at his approach, for he is a man of immense strength, and absolutely uncontrollable
in his anger.

"Last week he hurled the local blacksmith over a parapet into a stream, and it was only by paying
over all the money which I could gather together that I was able to avert another public exposure.
He had no friends at all save the wandering gypsies, and he would give these vagabonds leave to
encamp upon the few acres of bramble-covered land which represent the family estate, and would
accept in return the hospitality of their tents, wandering away with them sometimes for weeks on
end. He has a passion also for Indian animals, which are sent over to him by a correspondent, and he
has at this moment a cheetah and a baboon, which wander freely over his grounds and are feared by
the villagers almost as much as their master.