49 Ventnor Terrace
The year '87 furnished us with a long series of cases of greater or less interest, of which I retain
the records. Among my headings under this one twelve months I find an account of the adventure of
the Paradol Chamber, of the Amateur Mendicant Society, who held a luxurious club in the lower vault
of a furniture warehouse, of the facts connected with the loss of the British barque "Sophy
Anderson", of the singular adventures of the Grice Patersons in the island of Uffa, and finally of
the Camberwell poisoning case. In the latter, as may be remembered, Sherlock Holmes was able, by
winding up the dead man's watch, to prove that it had been wound up two hours before, and that
therefore the deceased had gone to bed within that time--a deduction which was of the greatest
importance in clearing up the case. All these I may sketch out at some future date, but none of them
present such singular features as the strange train of circumstances which I have now taken up my
pen to describe.

It was in the latter days of September, and the equinoctial gales had set in with exceptional
violence. All day the wind had screamed and the rain had beaten against the windows, so that even
here in the heart of great, hand-made London we were forced to raise our minds for the instant from
the routine of life and to recognise the presence of those great elemental forces which shriek at
mankind through the bars of his civilisation, like untamed beasts in a cage. As evening drew in, the
storm grew higher and louder, and the wind cried and sobbed like a child in the chimney. Sherlock
Holmes sat moodily at one side of the fireplace cross-indexing his records of crime, while I at the
other was deep in one of Clark Russell's fine sea-stories until the howl of the gale from without
seemed to blend with the text, and the splash of the rain to lengthen out into the long swash of the
sea waves. My wife was on a visit to her mother's, and for a few days I was a dweller once more in
my old quarters at Baker Street.

"Why," said I, glancing up at my companion, "that was surely the bell. Who could come to-night? Some
friend of yours, perhaps?"

"Except yourself I have none," he answered. "I do not encourage visitors."

"A client, then?"