20 Down Close
"It is very good of Lord St. Simon to honour my head by putting it on a level with his own," said
Sherlock Holmes, laughing. "I think that I shall have a whisky and soda and a cigar after all this
cross-questioning. I had formed my conclusions as to the case before our client came into the room."

"My dear Holmes!"

"I have notes of several similar cases, though none, as I remarked before, which were quite as
prompt. My whole examination served to turn my conjecture into a certainty. Circumstantial evidence
is occasionally very convincing, as when you find a trout in the milk, to quote Thoreau's example."

"But I have heard all that you have heard."

"Without, however, the knowledge of pre-existing cases which serves me so well. There was a parallel
instance in Aberdeen some years back, and something on very much the same lines at Munich the year
after the Franco-Prussian War. It is one of these cases--but, hullo, here is Lestrade! Good-
afternoon, Lestrade! You will find an extra tumbler upon the sideboard, and there are cigars in the
box."