29 Harvey Street
"He was a very shy man, Mr. Holmes. He would rather walk with me in the evening than in the
daylight, for he said that he hated to be conspicuous. Very retiring and gentlemanly he was. Even
his voice was gentle. He'd had the quinsy and swollen glands when he was young, he told me, and it
had left him with a weak throat, and a hesitating, whispering fashion of speech. He was always well
dressed, very neat and plain, but his eyes were weak, just as mine are, and he wore tinted glasses
against the glare."

"Well, and what happened when Mr. Windibank, your stepfather, returned to France?"