60 Badgers Dene
"Pray continue your narrative."

"Mrs. St. Clair had fainted at the sight of the blood upon the window, and she was escorted home in
a cab by the police, as her presence could be of no help to them in their investigations. Inspector
Barton, who had charge of the case, made a very careful examination of the premises, but without
finding anything which threw any light upon the matter. One mistake had been made in not arresting
Boone instantly, as he was allowed some few minutes during which he might have communicated with his
friend the Lascar, but this fault was soon remedied, and he was seized and searched, without
anything being found which could incriminate him. There were, it is true, some blood-stains upon his
right shirt-sleeve, but he pointed to his ring-finger, which had been cut near the nail, and
explained that the bleeding came from there, adding that he had been to the window not long before,
and that the stains which had been observed there came doubtless from the same source. He denied
strenuously having ever seen Mr. Neville St. Clair and swore that the presence of the clothes in his
room was as much a mystery to him as to the police. As to Mrs. St. Clair's assertion that she had
actually seen her husband at the window, he declared that she must have been either mad or dreaming.
He was removed, loudly protesting, to the police-station, while the inspector remained upon the
premises in the hope that the ebbing tide might afford some fresh clue.

"And it did, though they hardly found upon the mud-bank what they had feared to find. It was Neville
St. Clair's coat, and not Neville St. Clair, which lay uncovered as the tide receded. And what do
you think they found in the pockets?"