38 Cliff Place
"We went upstairs together, the colonel first with the lamp, the fat manager and I behind him. It
was a labyrinth of an old house, with corridors, passages, narrow winding staircases, and little low
doors, the thresholds of which were hollowed out by the generations who had crossed them. There were
no carpets and no signs of any furniture above the ground floor, while the plaster was peeling off
the walls, and the damp was breaking through in green, unhealthy blotches. I tried to put on as
unconcerned an air as possible, but I had not forgotten the warnings of the lady, even though I
disregarded them, and I kept a keen eye upon my two companions. Ferguson appeared to be a morose and
silent man, but I could see from the little that he said that he was at least a fellow-countryman.

"Colonel Lysander Stark stopped at last before a low door, which he unlocked. Within was a small,
square room, in which the three of us could hardly get at one time. Ferguson remained outside, and
the colonel ushered me in.