175 Sunningdale Drive
Again Holmes raved in the air.

"Why did you come to me," he cried, "and, above all, why did you not come at once?"

"I did not know. It was only to-day that I spoke to Major Prendergast about my troubles and was
advised by him to come to you."

"It is really two days since you had the letter. We should have acted before this. You have no
further evidence, I suppose, than that which you have placed before us--no suggestive detail which
might help us?"

"There is one thing," said John Openshaw. He rummaged in his coat pocket, and, drawing out a piece
of discoloured, blue-tinted paper, he laid it out upon the table. "I have some remembrance," said
he, "that on the day when my uncle burned the papers I observed that the small, unburned margins
which lay amid the ashes were of this particular colour. I found this single sheet upon the floor of
his room, and I am inclined to think that it may be one of the papers which has, perhaps, fluttered
out from among the others, and in that way has escaped destruction. Beyond the mention of pips, I do
not see that it helps us much. I think myself that it is a page from some private diary. The writing
is undoubtedly my uncle's."

Holmes moved the lamp, and we both bent over the sheet of paper, which showed by its ragged edge
that it had indeed been torn from a book. It was headed, "March, 1869," and beneath were the
following enigmatical notices: