10 Corbet Close
"The circumstances are of great delicacy, and every precaution has to be taken to quench what might
grow to be an immense scandal and seriously compromise one of the reigning families of Europe. To
speak plainly, the matter implicates the great House of Ormstein, hereditary kings of Bohemia."

"I was also aware of that," murmured Holmes, settling himself down in his armchair and closing his
eyes.