90 Holmefield Avenue
"And your father?" asked Holmes. "Was he in favour of such a union?"

"No, he was averse to it also. No one but Mr. McCarthy was in favour of it." A quick blush passed
over her fresh young face as Holmes shot one of his keen, questioning glances at her.

"Thank you for this information," said he. "May I see your father if I call to-morrow?"