40 Down Close
"I tell you that I would give one of the provinces of my kingdom to have that photograph."

"And for present expenses?"

The King took a heavy chamois leather bag from under his cloak and laid it on the table.

"There are three hundred pounds in gold and seven hundred in notes," he said.

Holmes scribbled a receipt upon a sheet of his note-book and handed it to him.