38 Holmefield Avenue
"No. If she can come to Winchester to meet us she can get away."

"Quite so. She has her freedom."

"What can be the matter, then? Can you suggest no explanation?"

"I have devised seven separate explanations, each of which would cover the facts as far as we know
them. But which of these is correct can only be determined by the fresh information which we shall
no doubt find waiting for us. Well, there is the tower of the cathedral, and we shall soon learn all
that Miss Hunter has to tell."