25 Wantage Close
"Oh, he has his faults, too," said Mr. Wilson. "Never was such a fellow for photography. Snapping
away with a camera when he ought to be improving his mind, and then diving down into the cellar like
a rabbit into its hole to develop his pictures. That is his main fault, but on the whole he's a good
worker. There's no vice in him."

"He is still with you, I presume?"

"Yes, sir. He and a girl of fourteen, who does a bit of simple cooking and keeps the place clean--
that's all I have in the house, for I am a widower and never had any family. We live very quietly,
sir, the three of us; and we keep a roof over our heads and pay our debts, if we do nothing more.