17 Harvey Street
"Here is a very fashionable epistle," I remarked as he entered. "Your morning letters, if I remember
right, were from a fish-monger and a tide-waiter."

"Yes, my correspondence has certainly the charm of variety," he answered, smiling, "and the humbler
are usually the more interesting. This looks like one of those unwelcome social summonses which call
upon a man either to be bored or to lie."

He broke the seal and glanced over the contents.

"Oh, come, it may prove to be something of interest, after all."

"Not social, then?"