131 Glan Road
"And she would need to be," said Holmes gravely. "I am much mistaken if we do not hear from her
before many days are past."

It was not very long before my friend's prediction was fulfilled. A fortnight went by, during which
I frequently found my thoughts turning in her direction and wondering what strange side-alley of
human experience this lonely woman had strayed into. The unusual salary, the curious conditions, the
light duties, all pointed to something abnormal, though whether a fad or a plot, or whether the man
were a philanthropist or a villain, it was quite beyond my powers to determine. As to Holmes, I
observed that he sat frequently for half an hour on end, with knitted brows and an abstracted air,
but he swept the matter away with a wave of his hand when I mentioned it. "Data! data! data!" he
cried impatiently. "I can't make bricks without clay." And yet he would always wind up by muttering
that no sister of his should ever have accepted such a situation.

The telegram which we eventually received came late one night just as I was thinking of turning in
and Holmes was settling down to one of those all-night chemical researches which he frequently
indulged in, when I would leave him stooping over a retort and a test-tube at night and find him in
the same position when I came down to breakfast in the morning. He opened the yellow envelope, and
then, glancing at the message, threw it across to me.

"Just look up the trains in Bradshaw," said he, and turned back to his chemical studies.

The summons was a brief and urgent one.

Please be at the Black Swan Hotel at Winchester at midday to-morrow [it said]. Do come! I am at my
wit's end. Hunter.