54 Glan Road
"You can easily imagine, Mr. Holmes, how curious I became as to what the meaning of this
extraordinary performance could possibly be. They were always very careful, I observed, to turn my
face away from the window, so that I became consumed with the desire to see what was going on behind
my back. At first it seemed to be impossible, but I soon devised a means. My hand-mirror had been
broken, so a happy thought seized me, and I concealed a piece of the glass in my handkerchief. On
the next occasion, in the midst of my laughter, I put my handkerchief up to my eyes, and was able
with a little management to see all that there was behind me. I confess that I was disappointed.
There was nothing. At least that was my first impression. At the second glance, however, I perceived
that there was a man standing in the Southampton Road, a small bearded man in a grey suit, who
seemed to be looking in my direction. The road is an important highway, and there are usually people
there. This man, however, was leaning against the railings which bordered our field and was looking
earnestly up. I lowered my handkerchief and glanced at Mrs. Rucastle to find her eyes fixed upon me
with a most searching gaze. She said nothing, but I am convinced that she had divined that I had a
mirror in my hand and had seen what was behind me. She rose at once.

"'Jephro,' said she, 'there is an impertinent fellow upon the road there who stares up at Miss
Hunter.'