83 Valentia Road
"That is well. And the Rucastles go out to-night?"

"Yes."

"Is there a cellar with a good strong lock?"

"Yes, the wine-cellar."

"You seem to me to have acted all through this matter like a very brave and sensible girl, Miss
Hunter. Do you think that you could perform one more feat? I should not ask it of you if I did not
think you a quite exceptional woman."

"I will try. What is it?"