38 Bowler Street
"How long I remained unconscious I cannot tell. It must have been a very long time, for the moon had
sunk, and a bright morning was breaking when I came to myself. My clothes were all sodden with dew,
and my coat-sleeve was drenched with blood from my wounded thumb. The smarting of it recalled in an
instant all the particulars of my night's adventure, and I sprang to my feet with the feeling that I
might hardly yet be safe from my pursuers. But to my astonishment, when I came to look round me,
neither house nor garden were to be seen. I had been lying in an angle of the hedge close by the
highroad, and just a little lower down was a long building, which proved, upon my approaching it, to
be the very station at which I had arrived upon the previous night. Were it not for the ugly wound
upon my hand, all that had passed during those dreadful hours might have been an evil dream.

"Half dazed, I went into the station and asked about the morning train. There would be one to
Reading in less than an hour. The same porter was on duty, I found, as had been there when I
arrived. I inquired of him whether he had ever heard of Colonel Lysander Stark. The name was strange
to him. Had he observed a carriage the night before waiting for me? No, he had not. Was there a
police-station anywhere near? There was one about three miles off.